+++ registration/src/main/java/ee/guest/registration/repositories/UserRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    Optional<User> findByPersonalCode(Integer personalCode);
+    Optional<User> findByPersonalCode(Long personalCode);
 
 }
